requirements.txt
tqdm
python-dotenv
google-generativeai
importlib-metadata
beautifulsoup4
readabilipy
trafilatura
datasets
nltk
numpy
requests
sentence-transformers
scikit-learn
sentence-transformers
scikit-learn